Address code review comments
- Rename test analyzers and improved AST-analysis logic.
- Use pass.Reportf in test analyzers.
- Documentation fixes.
- Add integer overflow check to analyzer-import generator.
- Removed global state from tests/core/nogo/deps/d.go.
